+++ app/_shared/note/services/noteActivityService/index.ts
@@ -1,0 +1,1 @@
+export * as noteActivityService from "./service";
