+++ app/_shared/note/services/noteCreateService/index.ts
@@ -1,0 +1,1 @@
+export * as noteCreateService from "./service";
