+++ app/_shared/note/services/noteFindService/index.ts
@@ -1,0 +1,1 @@
+export * as noteFindService from "./service";
