--- app/_shared/note/services/userFindService/index.ts
@@ -1,1 +1,0 @@
-export * as userFindService from "./service";
